requirements.txt
Flask>=2.0
python-dotenv>=0.21
openai>=1.0.0
flask-cors>=3.0.10
requests>=2.28
